requirements.txt
Django==4.0
django_filter==21.1
django_heroku==0.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pygments==2.10.0
python-decouple==3.5
psycopg2-binary==2.9.3
requests==2.27.0
